quiz

+++ backoffice/src/app/collaboration/collaborations/collaborations.component.ts
@@ -7,6 +7,7 @@ import { CollaborationService } from '../../Service/collaboration.service';
 import {NgxPaginationModule} from 'ngx-pagination';
 import { UpdateCollaborationComponent } from '../update-collaboration/update-collaboration.component';
 import {MatDialog, MAT_DIALOG_DATA} from '@angular/material/dialog';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-collaborations',
   templateUrl: './collaborations.component.html',
@@ -48,8 +49,42 @@ export class CollaborationsComponent implements OnInit {
 
 
       deleteCollaboration(idCollaboration:number){
-        this.collaborationService.deleteCollaboration(idCollaboration)
-        .subscribe(()=>this.collaborationService.getAllColaboration().subscribe(res=>{this.c=res}));
+        const swalWithBootstrapButtons = Swal.mixin({
+          customClass: {
+            confirmButton: 'btn btn-success',
+            cancelButton: 'btn btn-danger'
+          },
+          buttonsStyling: false
+        })
+        
+        swalWithBootstrapButtons.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonText: 'Yes, delete it!',
+          cancelButtonText: 'No, cancel!',
+          reverseButtons: true
+        }).then((result) => {
+          if (result.isConfirmed) {
+            this.collaborationService.deleteCollaboration(idCollaboration)
+            .subscribe(()=>this.collaborationService.getAllColaboration().subscribe(res=>{this.c=res}));
+            swalWithBootstrapButtons.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+          } else if (
+            /* Read more about handling dismissals below */
+            result.dismiss === Swal.DismissReason.cancel
+          ) {
+            swalWithBootstrapButtons.fire(
+              'Cancelled',
+              'Your imaginary file is safe :)',
+              'error'
+            )
+          }
+        })
       }
 
       updateCollaboration(idCollaboration:number){
