+++ backoffice/src/app/collaboration/offer/offer.component.ts
@@ -5,6 +5,7 @@ import { Offer } from './../../Models/Collaboration/offer';
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-offer',
@@ -38,8 +39,42 @@ export class OfferComponent implements OnInit {
   }
 
   deleteOffer(idOffer:number){
-    this.OfferService.deleteOffer(idOffer)
-    .subscribe(()=>this.OfferService.getAllOffer().subscribe(res=>{this.o=res}));
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false
+    })
+    
+    swalWithBootstrapButtons.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, cancel!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.OfferService.deleteOffer(idOffer)
+        .subscribe(()=>this.OfferService.getAllOffer().subscribe(res=>{this.o=res}));
+        swalWithBootstrapButtons.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          'Cancelled',
+          'Your imaginary file is safe :)',
+          'error'
+        )
+      }
+    })
   }
   updateOffer(idCollaboration:number){
